requirements.txt
fastapi[standard]==0.124.4
python-dotenv==1.2.2
pydantic==2.12.5
aiohttp==3.13.3
aiofiles==25.1.0
requests==2.32.5
pandas==3.0.1
xlrd >= 2.0.1
pymongo==4.16.0
beautifulsoup4==4.14.3
python-calamine==0.6.2
black==26.3.1
pre-commit==4.5.1
